+++ CHANGELOG.md
@@ -25,3 +25,6 @@ All notable changes to the "custom-search-filters" extension will be documented 
 ## [0.1.2] - 2025-04-21
 ### Enhanced
 - support add multiple folders and file types at once
+
+### Fixed
+- Fixed bug where edit filter command was not working.